add-pak as regular user, on Linux
