+++ src/macos.rs
@@ -255,57 +255,6 @@ pub fn sc_rm(args: &ArgMatches) -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
-pub fn system_add_pak(
-    vers: Option<Vec<String>>,
-    stream: &str,
-    update: bool,
-) -> Result<(), Box<dyn Error>> {
-    let vers = match vers {
-        Some(x) => x,
-        None => vec![sc_get_default_or_fail()?],
-    };
-
-    for ver in vers {
-        check_installed(&ver)?;
-        if update {
-            info!("Installing pak for R {}", ver);
-        } else {
-            info!("Installing pak for R {} (if not installed yet)", ver);
-        }
-        check_has_pak(&ver)?;
-
-        // We do this to create the user library, because currently there
-        // is a bug in the system profile code that creates it, and it is
-        // only added after a restart.
-        match r(&ver, "invisible()") {
-            Ok(_) => {},
-            Err(x) => bail!("Failed to to install pak for R {}: {}", ver, x.to_string())
-        };
-
-        // The actual pak installation
-        let cmd;
-        if update {
-            cmd = r#"
-                install.packages("pak", repos = sprintf("https://r-lib.github.io/p/pak/{}/%s/%s/%s", .Platform$pkgType, R.Version()$os, R.Version()$arch))
-            "#;
-        } else {
-            cmd = r#"
-                if (!requireNamespace("pak", quietly = TRUE)) {
-                    install.packages("pak", repos = sprintf("https://r-lib.github.io/p/pak/{}/%s/%s/%s", .Platform$pkgType, R.Version()$os, R.Version()$arch))
-                }
-            "#;
-        };
-        let cmd = cmd.replace("{}", stream);
-
-        match r(&ver, &cmd) {
-            Ok(_) => {},
-            Err(x) => bail!("Failed to install pak for R {}: {}", ver, x.to_string())
-        };
-    }
-
-    Ok(())
-}
-
 pub fn sc_system_make_links() -> Result<(), Box<dyn Error>> {
     escalate("making R-* quick links")?;
     let vers = sc_get_list()?;
